// src/types/getDiagnosesParams.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * Hopital Management System Api
 * OpenAPI spec version: v1
 */

export type GetDiagnosesParams = {
    'AppointmentId.Equal'?: number;
    'AppointmentId.NotEquals'?: number;
    'AppointmentId.GreaterThan'?: number;
    'AppointmentId.GreaterThanOrEqual'?: number;
    'AppointmentId.LessThan'?: number;
    'AppointmentId.LessThanOrEqual'?: number;
    'AppointmentId.In'?: number[];
    'AppointmentId.IsNull'?: boolean;
    'AppointmentId.IsNotNull'?: boolean;
    'DiagnosisDate.Equal'?: string;
    'DiagnosisDate.NotEquals'?: string;
    'DiagnosisDate.GreaterThan'?: string;
    'DiagnosisDate.GreaterThanOrEqual'?: string;
    'DiagnosisDate.LessThan'?: string;
    'DiagnosisDate.LessThanOrEqual'?: string;
    'DiagnosisDate.IsNull'?: boolean;
    'DiagnosisDate.IsNotNull'?: boolean;
    Search?: string;
    Page?: number;
    PageSize?: number;
    Skip?: number;
    Take?: number;
    SortBy?: string;
    SortOrder?: string;
};