// src/types/patient.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * Hopital Management System Api
 * OpenAPI spec version: v1
 */
import type { Appointment } from './appointment';
import type { EmergencyContact } from './emergencyContact';
import type { Invoice } from './invoice';
import type { IdentityUser } from './identityUser';

export interface Patient {
    /** @nullable */
    appointments?: Appointment[] | null;
    dateOfBirth?: string;
    /**
     * @minLength 0
     * @maxLength 100
     * @nullable
     */
    email?: string | null;
    /** @nullable */
    emergencyContacts?: EmergencyContact[] | null;
    /**
     * @minLength 0
     * @maxLength 10
     * @nullable
     */
    gender?: string | null;
    /**
     * @minLength 0
     * @maxLength 100
     * @nullable
     */
    healthInsurance?: string | null;
    /** @nullable */
    invoices?: Invoice[] | null;
    /**
     * @minLength 0
     * @maxLength 100
     * @nullable
     */
    name?: string | null;
    patientId?: number;
    /**
     * @minLength 0
     * @maxLength 15
     * @nullable
     */
    phoneNumber?: string | null;
    user?: IdentityUser;
    /** @nullable */
    userId?: string | null;
}